add missed build
