+++ efficientdet/keras/inference.py
@@ -113,13 +113,6 @@ def visualize_image(image,
   return img
 
 
-def get_ema_name(ema, var):
-  if var.ref() in ema._averages:
-    return ema._averages[var.ref()].name.split(':')[0]
-  return tf.compat.v1.get_default_graph().unique_name(
-      var.name.split(':')[0] + "/" + ema.name, mark_as_used=False)
-
-
 class ExportModel(tf.Module):
 
   def __init__(self, model):
@@ -220,37 +213,6 @@ class ServingDriver(object):
     policy = tf.keras.mixed_precision.experimental.Policy(precision)
     tf.keras.mixed_precision.experimental.set_policy(policy)
 
-  def restore_ckpt(self, ckpt_path, ema_decay=0.9998):
-    """Restore variables from a given checkpoint.
-
-    Args:
-      ckpt_path: the path of the checkpoint. Can be a file path or a folder path.
-      ema_decay: ema decay rate. If None or zero or negative value, disable ema.
-    """
-    if tf.io.gfile.isdir(ckpt_path):
-      ckpt_path = tf.train.latest_checkpoint(ckpt_path)
-    if ema_decay > 0:
-      ema = tf.train.ExponentialMovingAverage(decay=0.0)
-      ema_vars = util_keras.get_ema_vars(self.model)
-      var_dict = {
-          get_ema_name(ema, var): var for (ref, var) in ema_vars.items()
-      }
-    else:
-      ema_vars = util_keras.get_ema_vars(self.model)
-      var_dict = ema_vars
-    for v in self.model.variables:
-      if v.ref() not in ema_vars:
-        var_dict[v.name.split(':')[0]] = v
-    # add variables that not in var_dict
-    if ckpt_path == '_':
-      logging.info('Running test: do not load any ckpt.')
-      return
-    try:
-      for key, var in var_dict.items():
-        var.assign(tf.train.load_variable(ckpt_path, key))
-    except tf.errors.NotFoundError:
-      self.model.load_weights(ckpt_path)
-
   def build(self, params_override=None):
     """Build model and restore checkpoints."""
     params = copy.deepcopy(self.params)
@@ -260,7 +222,7 @@ class ServingDriver(object):
     config.override(params)
     self.model = EfficientDetModel(config=config)
     self.model.build((self.batch_size, *params['image_size'], 3))
-    self.restore_ckpt(self.ckpt_path, params['moving_average_decay'])
+    util_keras.restore_ckpt(self.model, self.ckpt_path, params['moving_average_decay'])
 
   def visualize(self, image, boxes, classes, scores, **kwargs):
     """Visualize prediction on image."""
